pyproject.toml
[tool.poetry]
name = "library-management"
version = "0.1.0"
description = ""
authors = ["MLEngineer12 <murjanivarun@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
setuptools = "<81"
python = ">=3.9,<4.0"
fastapi = {extras = ["all"], version = ">=0.116.1,<0.117.0"}
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
sqlalchemy = ">=2.0.42,<3.0.0"
pymysql = ">=1.1.1,<2.0.0"
alembic = ">=1.16.4,<2.0.0"
pydantic = ">=2.11.7,<3.0.0"

opentelemetry-api = {version = "==1.22.0", allow-prereleases = true}
opentelemetry-sdk = {version = "==1.22.0", allow-prereleases = true}
opentelemetry-exporter-otlp = {version = "==1.22.0", allow-prereleases = true}
opentelemetry-instrumentation = {version = "==0.43b0", allow-prereleases = true}
opentelemetry-instrumentation-fastapi = {version = "==0.43b0", allow-prereleases = true}
opentelemetry-instrumentation-logging = {version = "==0.43b0", allow-prereleases = true}
opentelemetry-semantic-conventions = {version = "==0.43b0", allow-prereleases = true}
honeycomb-opentelemetry = {version = "==0.6.0b0", allow-prereleases = true}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-dotenv = "^1.1.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest = "^8.4.2"
mysql-connector-python = "^9.4.0"
sudo = "^1.0.0"
debugpy = "^1.8.17"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
